requirements.txt
Flask
coverage
PyYAML